requirements.txt
Django==4.1
django-polymorphic==3.1.0
